LLM-VM/archive/binary-classification-version/requirements.txt
# Core dependencies
openai>=1.0.0
pyyaml>=6.0

# For embedding-based classifiers (entailment & cosine)
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0

# Optional: For development
pytest>=7.0.0

